requirements.txt
langchain==1.0.2
langchain-openai==1.0.1
langchain-mcp-adapters>=0.1.0
fastmcp==2.12.5
requests>=2.31.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
